Cleaned up a few more interfaces
