+++ src/autoMutator.ts
@@ -1,6 +1,6 @@
 import { Logger } from "./logger";
 import { ConsoleLogger } from "./loggers/consoleLogger";
-import { MutationsApplier } from "./mutationsApplier";
+import { MutationsApplier } from "./types/mutationsApplier";
 import { FileMutationsApplier } from "./mutationsAppliers/fileMutationsApplier";
 import { MutationsProvider, MutationsWave } from "./mutationsProvider";
 import { MutationRunSettings } from "./runMutations";
