+++ src/nameTransformer.ts
@@ -1,62 +1,36 @@
 /**
- * Transforms dashed-case names to cases.
+ * Transforms a dashed-case name to camelCase.
+ *
+ * @param name   A dashed-case name.
+ * @returns The name as camelCase.
  */
-export interface NameTransformer {
-  /**
-   * Transforms a dashed-case name to camelCase.
-   *
-   * @param name   A dashed-case name.
-   * @returns The name as camelCase.
-   */
-  toCamelCase(name: string): string;
+export const toCamelCase = (name: string): string => {
+  const split: string[] = name.split("-");
 
-  /**
-   * Transforms a dashed-case name to PamelCase.
-   *
-   * @param name   A dashed-case name.
-   * @returns The name as PascalCase.
-   */
-  toPascalCase(name: string): string;
-}
-
-/**
- * Transforms dashed-case names to cases.
- */
-export class NameTransformer implements NameTransformer {
-  /**
-   * Transforms a dashed-case name to camelCase.
-   *
-   * @param name   A dashed-case name.
-   * @returns The name as camelCase.
-   */
-  public toCamelCase(name: string): string {
-    const split: string[] = name.split("-");
-
-    return (
-      split[0].toLowerCase() +
-      split
-        .slice(1)
-        .map(
-          (part: string) =>
-            part.substring(0, 1).toUpperCase() + part.substring(1).toLowerCase()
-        )
-        .join("")
-    );
-  }
-
-  /**
-   * Transforms a dashed-case name to PamelCase.
-   *
-   * @param name   A dashed-case name.
-   * @returns The name as PascalCase.
-   */
-  public toPascalCase(name: string): string {
-    return name
-      .split("-")
+  return (
+    split[0].toLowerCase() +
+    split
+      .slice(1)
       .map(
         (part: string) =>
           part.substring(0, 1).toUpperCase() + part.substring(1).toLowerCase()
       )
-      .join("");
-  }
-}
+      .join("")
+  );
+};
+
+/**
+ * Transforms a dashed-case name to PascalCase.
+ *
+ * @param name   A dashed-case name.
+ * @returns The name as PascalCase.
+ */
+export const toPascalCase = (name: string): string => {
+  return name
+    .split("-")
+    .map(
+      (part: string) =>
+        part.substring(0, 1).toUpperCase() + part.substring(1).toLowerCase()
+    )
+    .join("");
+};
